requirements.txt
langchain>=0.1.0
langgraph>=0.0.30
langchain-openai>=0.1.0
langchain-ollama>=0.0.6
python-dotenv>=1.0.0
openai>=1.0.0
tqdm>=4.0.0
typing-extensions>=4.0.0